requirements.txt
tensorflow
numpy
pandas
matplotlib
scikit-learn
Pillow

#pip install -r requirements.txt